+++ src/bookshelf/reading-journey/position/position.spec.ts
@@ -3,6 +3,8 @@ import { position } from './position'
 import { Page } from './page'
 import { Percentage } from './percentage'
 import { RomanNumeral } from './roman-numeral'
+import { Time } from './time'
+import { Playtime } from '../../shared/playtime'
 
 test.each([
     [1, new Page(1)],
@@ -10,6 +12,7 @@ test.each([
     [' 123 ', new Page(123)],
     ['25%', new Percentage(25)],
     ['xxiv', new RomanNumeral(24)],
+    ['10:35', new Time(Playtime.fromMinutes(635))],
 ])('position(%o) should return %o', (input, expected) => {
     expect(position(input)).toEqual(expected)
 })